+++ VS_SolarAsh/Assets/Scenes/Attack.unity
@@ -630,6 +630,18 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   ani: {fileID: 774531905}
+--- !u!114 &774531908
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 774531900}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: e9c97a0772a9f2949b1ad09383b61644, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
   targetTag: Boxes
 --- !u!1 &1454180122
 GameObject:
@@ -1427,6 +1439,10 @@ PrefabInstance:
         type: 3}
       insertIndex: -1
       addedObject: {fileID: 774531907}
+    - targetCorrespondingSourceObject: {fileID: 919132149155446097, guid: 7d32e8de2e56b9c49b9cd519a6683a52,
+        type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 774531908}
     - targetCorrespondingSourceObject: {fileID: -5195048882033904264, guid: 7d32e8de2e56b9c49b9cd519a6683a52,
         type: 3}
       insertIndex: -1
