+++ VS_SolarAsh/Assets/script/player/AttackHit.cs
@@ -1,0 +1,16 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class AttackHit : MonoBehaviour
+{
+    [SerializeField] private string targetTag;
+
+    private void OnTriggerEnter(Collider other)
+    {
+        if (other.gameObject.tag == targetTag)
+        {
+            GameObject.Destroy(other.gameObject);
+        }
+    }
+}
