+++ VS_SolarAsh/Assets/script/player/AttackHit.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e9c97a0772a9f2949b1ad09383b61644
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
